+++ packages/synchro-charts/cypress/integration/charts/bar-chart/tooltip.spec.ts
@@ -78,8 +78,6 @@ it('renders tooltip rows in order of values magnitude', () => {
     offsetY: SCREEN_SIZE.height / 2,
   });
 
-  cy.wait(0.05 * SECOND_IN_MS);
-
   cy.get(CHART_TOOLTIP_SELECTOR).should('be.visible');
   cy.get(CHART_TOOLTIP_ROW_SELECTOR).should('have.length', 2);
 
@@ -102,6 +100,26 @@ it('renders tooltip rows in order of values magnitude', () => {
     .eq(1)
     .contains(NUMBER_STREAM_1.data[0].y)
     .should('be.visible');
+
+  cy.get(CHART_VIZ_CONTAINER_SELECTOR).matchImageSnapshotOnCI();
+});
+
+it('renders tooltip to the left of the mouse when the mouse is on the right side', () => {
+  visitDynamicWidget(cy, {
+    componentTag: 'sc-bar-chart',
+    viewportStart: new Date(new Date(2000, 0, 0).getTime() - MINUTE_IN_MS),
+    viewportEnd: new Date(2000, 0, 0, 0, 5),
+    dataStreams: [NUMBER_STREAM_1],
+  });
+
+  cy.waitForChart();
+
+  cy.get(CHART_VIZ_CONTAINER_SELECTOR).trigger('mousemove', {
+    offsetX: 300,
+    offsetY: SCREEN_SIZE.height / 2,
+  });
+
+  cy.get(CHART_TOOLTIP_SELECTOR).should('be.visible');
 
   cy.get(CHART_VIZ_CONTAINER_SELECTOR).matchImageSnapshotOnCI();
 });